+++ session01/statistics/min.py
@@ -1,8 +1,8 @@
 numstr = input("to find min, input a set of numbers separated by space: ")
 try:
     numlist = numstr.split( )
-    numint = map(int, numlist)
-    numint = list(map(int, numlist))#conv to int
+    numint = map(float, numlist)
+    numint = list(map(float, numlist))#conv to int
     numint.sort()#min to max
     print(numint[0])
 except ValueError:#not num or separated by comma
